Output finite state machine changes in DEBUG mode Only transition states if they actually change
Add a battery LED that shows green if "healthy," yellow if low and red if critical.
Add logic to detect if we are running via USB or the battery.
If powered by the battery, added logic to automatically disable Ada and enter deep sleep mode to prevent damaging a critically low battery.